+++ app/lib/mercator.py
@@ -10,8 +10,8 @@ else:
 
 def mercator(
     coords: NDArray[np.floating],
-    width: cython.int,
-    height: cython.int,
+    width: cython.size_t,
+    height: cython.size_t,
 ) -> NDArray[np.floating]:
     xs = coords[:, 0]
     ys = _y_sheet(coords[:, 1])
