+++ app/middlewares/unsupported_browser_middleware.py
@@ -63,7 +63,7 @@ class UnsupportedBrowserMiddleware:
 
 @cython.cfunc
 def _should_capture(message: Message) -> cython.bint:
-    status_code: cython.int = message['status']
+    status_code: cython.size_t = message['status']
     if status_code < 200 or status_code >= 300:
         return False
